RavenDB-25281 -  Refactor RoaringBitmap implementation
