+++ test/FastTests/Corax/RoaringBitmapTests.cs
@@ -611,133 +611,6 @@ public unsafe class RoaringBitmapTests : NoDisposalNeeded
 
     #endregion
 
-    #region SIMD Correctness Verification
-
-    [RavenFact(RavenTestCategory.Corax)]
-    public unsafe void SimdAndMatchesScalar()
-    {
-        using var ctx = new ByteStringContext(SharedMultipleUseFlag.None);
-        int count = RoaringBitmap.BitmapContainerSizeInUInt64;
-
-        ctx.Allocate(RoaringBitmap.BitmapContainerSizeInBytes, out var aStorage);
-        ctx.Allocate(RoaringBitmap.BitmapContainerSizeInBytes, out var bStorage);
-        ctx.Allocate(RoaringBitmap.BitmapContainerSizeInBytes, out var simdResult);
-        ctx.Allocate(RoaringBitmap.BitmapContainerSizeInBytes, out var scalarResult);
-
-        try
-        {
-            ulong* a = (ulong*)aStorage.Ptr;
-            ulong* b = (ulong*)bStorage.Ptr;
-            ulong* sr = (ulong*)simdResult.Ptr;
-            ulong* sc = (ulong*)scalarResult.Ptr;
-
-            Random rng = new(42);
-            for (int i = 0; i < count; i++)
-            {
-                a[i] = (ulong)rng.NextInt64();
-                b[i] = (ulong)rng.NextInt64();
-            }
-
-            int simdCard = RoaringBitmap.BitmapAndSimd(a, b, sr, count);
-            int scalarCard = RoaringBitmap.BitmapAndScalar(a, b, sc, count);
-
-            Assert.Equal(scalarCard, simdCard);
-            for (int i = 0; i < count; i++)
-                Assert.Equal(sc[i], sr[i]);
-        }
-        finally
-        {
-            ctx.Release(ref aStorage);
-            ctx.Release(ref bStorage);
-            ctx.Release(ref simdResult);
-            ctx.Release(ref scalarResult);
-        }
-    }
-
-    [RavenFact(RavenTestCategory.Corax)]
-    public unsafe void SimdOrMatchesScalar()
-    {
-        using var ctx = new ByteStringContext(SharedMultipleUseFlag.None);
-        int count = RoaringBitmap.BitmapContainerSizeInUInt64;
-
-        ctx.Allocate(RoaringBitmap.BitmapContainerSizeInBytes, out var aStorage);
-        ctx.Allocate(RoaringBitmap.BitmapContainerSizeInBytes, out var bStorage);
-        ctx.Allocate(RoaringBitmap.BitmapContainerSizeInBytes, out var simdResult);
-        ctx.Allocate(RoaringBitmap.BitmapContainerSizeInBytes, out var scalarResult);
-
-        try
-        {
-            ulong* a = (ulong*)aStorage.Ptr;
-            ulong* b = (ulong*)bStorage.Ptr;
-            ulong* sr = (ulong*)simdResult.Ptr;
-            ulong* sc = (ulong*)scalarResult.Ptr;
-
-            Random rng = new(42);
-            for (int i = 0; i < count; i++)
-            {
-                a[i] = (ulong)rng.NextInt64();
-                b[i] = (ulong)rng.NextInt64();
-            }
-
-            int simdCard = RoaringBitmap.BitmapOrSimd(a, b, sr, count);
-            int scalarCard = RoaringBitmap.BitmapOrScalar(a, b, sc, count);
-
-            Assert.Equal(scalarCard, simdCard);
-            for (int i = 0; i < count; i++)
-                Assert.Equal(sc[i], sr[i]);
-        }
-        finally
-        {
-            ctx.Release(ref aStorage);
-            ctx.Release(ref bStorage);
-            ctx.Release(ref simdResult);
-            ctx.Release(ref scalarResult);
-        }
-    }
-
-    [RavenFact(RavenTestCategory.Corax)]
-    public unsafe void SimdAndNotMatchesScalar()
-    {
-        using var ctx = new ByteStringContext(SharedMultipleUseFlag.None);
-        int count = RoaringBitmap.BitmapContainerSizeInUInt64;
-
-        ctx.Allocate(RoaringBitmap.BitmapContainerSizeInBytes, out var aStorage);
-        ctx.Allocate(RoaringBitmap.BitmapContainerSizeInBytes, out var bStorage);
-        ctx.Allocate(RoaringBitmap.BitmapContainerSizeInBytes, out var simdResult);
-        ctx.Allocate(RoaringBitmap.BitmapContainerSizeInBytes, out var scalarResult);
-
-        try
-        {
-            ulong* a = (ulong*)aStorage.Ptr;
-            ulong* b = (ulong*)bStorage.Ptr;
-            ulong* sr = (ulong*)simdResult.Ptr;
-            ulong* sc = (ulong*)scalarResult.Ptr;
-
-            Random rng = new(42);
-            for (int i = 0; i < count; i++)
-            {
-                a[i] = (ulong)rng.NextInt64();
-                b[i] = (ulong)rng.NextInt64();
-            }
-
-            int simdCard = RoaringBitmap.BitmapAndNotSimd(a, b, sr, count);
-            int scalarCard = RoaringBitmap.BitmapAndNotScalar(a, b, sc, count);
-
-            Assert.Equal(scalarCard, simdCard);
-            for (int i = 0; i < count; i++)
-                Assert.Equal(sc[i], sr[i]);
-        }
-        finally
-        {
-            ctx.Release(ref aStorage);
-            ctx.Release(ref bStorage);
-            ctx.Release(ref simdResult);
-            ctx.Release(ref scalarResult);
-        }
-    }
-
-    #endregion
-
     #region Stress / Large Scale Tests
 
     [RavenFact(RavenTestCategory.Corax)]
